requirements.txt
bcrypt==3.2.0
black==21.9b0
blinker==1.4
cffi==1.15.0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pycparser==2.20
PyJWT==2.3.0
regex==2021.10.23
six==1.16.0
SQLAlchemy==1.4.26
tomli==1.2.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
WTForms==2.3.3
